svelte-dialog(refactor): use updated state model
