+++ packages/common/svelte/lib/main.ts
@@ -3,5 +3,5 @@ export * from './createContext';
 export * from './createEventForwarder';
 export * from './createRefAction';
 export * from './createSyncedOption';
-export * from './slots';
 export * from './store';
+export * from './types';
