+++ packages/dialog/svelte/lib/DialogTrigger.svelte
@@ -43,7 +43,7 @@
 	);
 	const id = component.getId();
 
-	const rootState = getDialogRootState() ?? readable(rootModel.getState());
+	const rootState = getDialogRootState() ?? readable(rootModel.state);
 
 	onMount(() => {
 		rootModel.mountComponent(id);
